requirements.txt
openai
streamlit
python-dotenv
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai  # Replace with the correct version if different
numpy
